# civicswipe/docs/api-spec.yaml
openapi: 3.1.0
info:
  title: CivicSwipe API
  version: 1.0.0
  description: |
    API for CivicSwipe - A legislation tracking and voting app.
    Phoenix, Arizona MVP Release.
    
    ## Authentication
    All endpoints except /auth/* require Bearer JWT token:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Pagination
    Cursor-based pagination using `cursor` and `limit` query parameters.
    
    ## Idempotency
    POST endpoints that record votes support `Idempotency-Key` header.
servers:
  - url: https://api.civicswipe.app/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local development
...
paths:
...
  /feed:
    get:
      tags: [Feed]
      summary: Get personalized swipe feed
      security:
        - BearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: level
          in: query
          schema:
            type: string
            enum: [federal, state, city, county]
        - name: status
          in: query
          schema:
            type: string
            enum: [introduced, scheduled, passed, failed]
            default: scheduled
        - name: topic
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Feed items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeedCard'
                  next_cursor:
                    type: string
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /measures/{measure_id}:
    get:
      tags: [Feed]
      summary: Get measure details
      security:
        - BearerAuth: []
      parameters:
        - name: measure_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Measure details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasureDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FeedCard:
      type: object
      properties:
        measure_id:
          type: string
          format: uuid
        title:
          type: string
        level:
          type: string
          enum: [federal, state, county, city]
        jurisdiction_name:
          type: string
        status:
          type: string
        scheduled_for:
          type: string
          format: date-time
        topic_tags:
          type: array
          items:
            type: string
        summary_short:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        user_vote:
          type: string
          enum: [yes, no]
          nullable: true
    Source:
      type: object
      properties:
        label:
          type: string
        url:
          type: string
          format: uri
        ctype:
          type: string
          enum: [html, pdf, api, text]
    MeasureDetail:
      type: object
      properties:
        measure:
          type: object
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
            level:
              type: string
            status:
              type: string
            introduced_at:
              type: string
              format: date-time
            scheduled_for:
              type: string
              format: date-time
            summary_short:
              type: string
            summary_long:
              type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        timeline:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              effective_at:
                type: string
                format: date-time
        vote_events:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              body:
                type: string
              scheduled_for:
                type: string
                format: date-time
              held_at:
                type: string
                format: date-time
                nullable: true
              result:
                type: string
        user_vote:
          type: object
          nullable: true
          properties:
            vote:
              type: string
            created_at:
              type: string
              format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...